plugins/skills/orchestrate-auto-dev/requirements.txt
psutil>=5.9.0
pydantic>=2.5.0
pyyaml>=6.0.1
